The solution to the best bags problem
There are n piles with each pile having n bags each. 
We are given a set of piles and we need to choose the bags  from the piles, 
such that
The number of bags  picked from ith pile  should be less than the number of bags  picked from the  i+1 th pile. 
get the total number of bags which  can be picked up from a given set of piles.